--- a/lazycache_test.go
+++ b/lazycache_test.go
@@ -1,6 +1,7 @@
 package lazycache
 
 import (
+	"errors"
 	"fmt"
 	"math/rand"
 	"sync"
@@ -14,65 +15,147 @@ import (
 func TestCache(t *testing.T) {
 	c := qt.New(t)
 
-	cache := New(CacheOptions{MaxEntries: 10})
+	cache := New[int, any](Options{MaxEntries: 10})
 
-	c.Assert(cache.Get("foo").Value(), qt.IsNil)
+	get := func(key int) any {
+		v, found := cache.Get(key)
+		if !found {
+			return nil
+		}
+		return v
+	}
 
-	cache.Set("foo", 32)
-	c.Assert(cache.Get("foo").Value(), qt.Equals, 32)
+	c.Assert(get(123456), qt.IsNil)
+
+	cache.Set(123456, 32)
+	c.Assert(get(123456), qt.Equals, 32)
 
 	for i := 0; i < 20; i++ {
 		cache.Set(i, i)
 	}
-	c.Assert(cache.Get("foo").Value(), qt.IsNil)
+	c.Assert(get(123456), qt.IsNil)
 	c.Assert(cache.Resize(5), qt.Equals, 5)
-	c.Assert(cache.Get(3).Value(), qt.IsNil)
-	c.Assert(cache.Contains(18), qt.IsTrue)
-	c.Assert(cache.Keys(), qt.DeepEquals, []any{15, 16, 17, 18, 19})
+	c.Assert(get(3), qt.IsNil)
+	c.Assert(cache.contains(18), qt.IsTrue)
+	c.Assert(cache.keys(), qt.DeepEquals, []int{15, 16, 17, 18, 19})
 
 	c.Assert(cache.DeleteFunc(
-		func(key any, value Entry) bool {
-			return value.Value().(int) > 15
+		func(key int, value any) bool {
+			return value.(int) > 15
 		},
 	), qt.Equals, 4)
 
-	c.Assert(cache.Contains(18), qt.IsFalse)
-	c.Assert(cache.Contains(15), qt.IsTrue)
+	c.Assert(cache.contains(18), qt.IsFalse)
+	c.Assert(cache.contains(15), qt.IsTrue)
 
 	c.Assert(cache.Delete(15), qt.IsTrue)
 	c.Assert(cache.Delete(15), qt.IsFalse)
-	c.Assert(cache.Contains(15), qt.IsFalse)
+	c.Assert(cache.contains(15), qt.IsFalse)
 
-	c.Assert(cache.Len(), qt.Equals, 0)
+	c.Assert(cache.len(), qt.Equals, 0)
 
-	c.Assert(func() { New(CacheOptions{MaxEntries: -1}) }, qt.PanicMatches, "Must provide a positive size")
+	c.Assert(func() { New[int, any](Options{MaxEntries: -1}) }, qt.PanicMatches, "must provide a positive size")
+
+}
+
+func TestDeleteFunc(t *testing.T) {
+	c := qt.New(t)
+
+	c.Run("Basic", func(c *qt.C) {
+		cache := New[int, any](Options{MaxEntries: 1000})
+		for i := 0; i < 10; i++ {
+			cache.Set(i, i)
+		}
+		c.Assert(cache.DeleteFunc(func(key int, value any) bool {
+			return key%2 == 0
+		}), qt.Equals, 5)
+		c.Assert(cache.len(), qt.Equals, 5)
+	})
+
+	c.Run("Temporary", func(c *qt.C) {
+
+		var wg sync.WaitGroup
+
+		// There's some timing involved in this test, so we'll need
+		// to retry a few times to cover all the cases.
+		for i := 0; i < 100; i++ {
+			cache := New[int, any](Options{MaxEntries: 1000})
+			for i := 0; i < 10; i++ {
+				cache.Set(i, i)
+			}
+			wg.Add(1)
+			go func() {
+				defer wg.Done()
+				for i := 10; i < 30; i++ {
+					v, _, err := cache.GetOrCreate(i, func(key int) (any, error) {
+						if key%2 == 0 {
+							return nil, errors.New("failed")
+						}
+						time.Sleep(10 * time.Microsecond)
+
+						return key, nil
+					})
+					if err != nil {
+						c.Assert(err, qt.ErrorMatches, "failed")
+					} else {
+						c.Assert(v, qt.Equals, i)
+					}
+
+				}
+			}()
+			time.Sleep(3 * time.Microsecond)
+			c.Assert(cache.DeleteFunc(func(key int, value any) bool {
+				return key%2 == 0
+			}), qt.Equals, 5)
+
+		}
+
+		wg.Wait()
+	})
 
 }
 
 func TestGetOrCreate(t *testing.T) {
 	c := qt.New(t)
-	cache := New(CacheOptions{MaxEntries: 100})
+	cache := New[int, any](Options{MaxEntries: 100})
 	counter := 0
-	create := func(key any) (any, error) {
+	create := func(key int) (any, error) {
 		counter++
-		return fmt.Sprintf("value-%s-%d", key, counter), nil
+		return fmt.Sprintf("value-%d-%d", key, counter), nil
 	}
 	for i := 0; i < 3; i++ {
-		res := cache.GetOrCreate("foo", create)
-		c.Assert(res.Err(), qt.IsNil)
-		c.Assert(res.Value(), qt.Equals, "value-foo-1")
+		res, found, err := cache.GetOrCreate(123456, create)
+		c.Assert(found, qt.IsTrue, qt.Commentf("iteration %d", i))
+		c.Assert(err, qt.IsNil)
+		c.Assert(res, qt.Equals, "value-123456-1")
 	}
-	c.Assert(cache.Get("foo").Value(), qt.Equals, "value-foo-1")
+	v, found := cache.Get(123456)
+	c.Assert(found, qt.IsTrue)
+	c.Assert(v, qt.Equals, "value-123456-1")
+}
+
+func TestGetOrCreateError(t *testing.T) {
+	c := qt.New(t)
+	cache := New[int, any](Options{MaxEntries: 100})
+	create := func(key int) (any, error) {
+		return nil, fmt.Errorf("failed")
+	}
+
+	res, found, err := cache.GetOrCreate(123456, create)
+	c.Assert(found, qt.IsFalse)
+	c.Assert(err, qt.ErrorMatches, "failed")
+	c.Assert(res, qt.IsNil)
+
 }
 
 func TestGetOrCreateConcurrent(t *testing.T) {
 	c := qt.New(t)
 
-	cache := New(CacheOptions{MaxEntries: 1000})
+	cache := New[int, any](Options{MaxEntries: 1000})
 
 	var countersmu sync.Mutex
-	counters := make(map[any]int)
-	create := func(key any) (any, error) {
+	counters := make(map[int]int)
+	create := func(key int) (any, error) {
 		countersmu.Lock()
 		count := counters[key]
 		counters[key]++
@@ -90,9 +173,10 @@ func TestGetOrCreateConcurrent(t *testing.T) {
 		go func() {
 			defer wg.Done()
 			for j := 0; j < 12; j++ {
-				res := cache.GetOrCreate(i, create)
-				c.Assert(res.Err(), qt.IsNil)
-				c.Assert(res.Value(), qt.Equals, expect)
+				res, found, err := cache.GetOrCreate(i, create)
+				c.Assert(found, qt.IsTrue)
+				c.Assert(err, qt.IsNil)
+				c.Assert(res, qt.Equals, expect)
 			}
 		}()
 	}
@@ -104,11 +188,10 @@ func TestGetOrCreateConcurrent(t *testing.T) {
 		go func() {
 			defer wg.Done()
 			for j := 0; j < 12; j++ {
-				res := cache.Get(i)
-				c.Assert(res.Err(), qt.IsNil)
+				res, found := cache.Get(i)
 				// The value may be nil if if GetOrCreate has not been called for i yet.
-				if v := res.Value(); v != nil {
-					c.Assert(v, qt.Equals, expect)
+				if found {
+					c.Assert(res, qt.Equals, expect)
 				}
 			}
 		}()
@@ -120,7 +203,7 @@ func TestGetOrCreateConcurrent(t *testing.T) {
 func BenchmarkGetOrCreate(b *testing.B) {
 	const maxSize = 1000
 
-	runBenchmark := func(b *testing.B, cache *Cache, getOrCreate func(key any, create func(key any) (any, error)) Entry) {
+	runBenchmark := func(b *testing.B, cache *Cache[int, any], getOrCreate func(key int, create func(key int) (any, error)) (any, bool, error)) {
 		r := rand.New(rand.NewSource(99))
 		var mu sync.Mutex
 
@@ -135,15 +218,12 @@ func BenchmarkGetOrCreate(b *testing.B) {
 				i1, i2 := r.Intn(maxSize), r.Intn(maxSize)
 				mu.Unlock()
 				// Just Get the value.
-				res1 := cache.Get(i1)
-				if err := res1.Err(); err != nil {
-					b.Fatal(err)
-				}
-				if v := res1.Value(); v != nil && v != i1 {
+				v, found := cache.Get(i1)
+				if found && v != i1 {
 					b.Fatalf("got %v, want %v", v, i1)
 				}
 
-				res2 := getOrCreate(i2, func(key any) (any, error) {
+				res2, found, err := getOrCreate(i2, func(key int) (any, error) {
 					if i2%100 == 0 {
 						// Simulate a slow create.
 						time.Sleep(1 * time.Second)
@@ -151,29 +231,19 @@ func BenchmarkGetOrCreate(b *testing.B) {
 					return i2, nil
 				})
 
-				if err := res2.Err(); err != nil {
+				if err != nil {
 					b.Fatal(err)
 				}
 
-				if v := res2.Value(); v != i2 {
+				if v := res2; !found || v != i2 {
 					b.Fatalf("got %v, want %v", v, i2)
 				}
 			}
 		})
 	}
 
-	/*b.Run("BaselineLock", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
-		runBenchmark(b, cache, cache.getOrCreateBaselineLock)
-	})
-
-	b.Run("BaselineDoubleCheckedLocking", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
-		runBenchmark(b, cache, cache.getOrCreateBaselDoubleCheckedLock)
-	})*/
-
 	b.Run("Real", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
+		cache := New[int, any](Options{MaxEntries: maxSize})
 		runBenchmark(b, cache, cache.GetOrCreate)
 	})
 
@@ -183,14 +253,14 @@ func BenchmarkCacheSerial(b *testing.B) {
 	const maxSize = 1000
 
 	b.Run("Set", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
+		cache := New[int, any](Options{MaxEntries: maxSize})
 		for i := 0; i < b.N; i++ {
 			cache.Set(i, i)
 		}
 	})
 
 	b.Run("Get", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
+		cache := New[int, any](Options{MaxEntries: maxSize})
 		numItems := maxSize - 200
 		for i := 0; i < numItems; i++ {
 			cache.Set(i, i)
@@ -198,8 +268,8 @@ func BenchmarkCacheSerial(b *testing.B) {
 		b.ResetTimer()
 		for i := 0; i < b.N; i++ {
 			key := i % numItems
-			v := cache.Get(key)
-			if v == nil {
+			_, found := cache.Get(key)
+			if !found {
 				b.Fatalf("unexpected nil value for key %d", key)
 			}
 		}
@@ -210,7 +280,7 @@ func BenchmarkCacheParallel(b *testing.B) {
 	const maxSize = 1000
 
 	b.Run("Set", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
+		cache := New[int, any](Options{MaxEntries: maxSize})
 		var counter uint32
 		b.RunParallel(func(pb *testing.PB) {
 			for pb.Next() {
@@ -221,7 +291,7 @@ func BenchmarkCacheParallel(b *testing.B) {
 	})
 
 	b.Run("Get", func(b *testing.B) {
-		cache := New(CacheOptions{MaxEntries: maxSize})
+		cache := New[int, any](Options{MaxEntries: maxSize})
 		r := rand.New(rand.NewSource(99))
 		var mu sync.Mutex
 		numItems := maxSize - 200
@@ -233,66 +303,11 @@ func BenchmarkCacheParallel(b *testing.B) {
 				mu.Lock()
 				key := r.Intn(numItems)
 				mu.Unlock()
-				v := cache.Get(key)
-				if v == nil {
+				_, found := cache.Get(key)
+				if !found {
 					b.Fatalf("unexpected nil value for key %d", key)
 				}
 			}
 		})
 	})
-}
-
-// These are only used in benchmarks.
-// This should be functionally equivalent to GetOrCreate.
-func (c *Cache) getOrCreateBaselineLock(key any, create func(key any) (any, error)) Entry {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	v, ok := c.lru.Get(key)
-	if ok {
-		return v.(Entry)
-	}
-
-	v, err := create(key)
-
-	e := entry{
-		value: v,
-		err:   err,
-	}
-
-	c.lru.Add(key, e)
-
-	return e
-
-}
-
-// This variant does not hold any lock while calling create, which means it may be called multiple times for the same key.
-func (c *Cache) getOrCreateBaselDoubleCheckedLock(key any, create func(key any) (any, error)) Entry {
-	c.mu.Lock()
-	v, ok := c.lru.Get(key)
-	if ok {
-		c.mu.Unlock()
-		return v.(Entry)
-	}
-	c.mu.Unlock()
-
-	v, err := create(key)
-
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	// Double check.
-	v2, ok := c.lru.Get(key)
-	if ok {
-		return v2.(Entry)
-	}
-
-	e := entry{
-		value: v,
-		err:   err,
-	}
-
-	c.lru.Add(key, e)
-
-	return e
-
 }