Which makes it more pleasant to use and, in general, faster:

```bash
name                 old time/op    new time/op    delta
GetOrCreate/Real-4      276µs ±82%     351µs ±43%     ~     (p=0.343 n=4+4)
CacheSerial/Set-4       176ns ± 0%     120ns ± 0%  -31.96%  (p=0.029 n=4+4)
CacheSerial/Get-4      27.1ns ± 2%    20.7ns ± 1%  -23.62%  (p=0.029 n=4+4)
CacheParallel/Set-4     272ns ± 1%     233ns ± 0%  -14.05%  (p=0.029 n=4+4)
CacheParallel/Get-4     149ns ± 1%     144ns ± 2%   -3.33%  (p=0.029 n=4+4)

name                 old alloc/op   new alloc/op   delta
GetOrCreate/Real-4      37.5B ± 9%     26.5B ± 9%  -29.33%  (p=0.029 n=4+4)
CacheSerial/Set-4        128B ± 0%      104B ± 0%  -18.75%  (p=0.029 n=4+4)
CacheSerial/Get-4       0.00B          0.00B          ~     (all equal)
CacheParallel/Set-4      128B ± 0%      104B ± 0%  -18.75%  (p=0.029 n=4+4)
CacheParallel/Get-4     0.00B          0.00B          ~     (all equal)

name                 old allocs/op  new allocs/op  delta
GetOrCreate/Real-4       1.50 ±33%      1.00 ± 0%     ~     (p=0.429 n=4+4)
CacheSerial/Set-4        4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.029 n=4+4)
CacheSerial/Get-4        0.00           0.00          ~     (all equal)
CacheParallel/Set-4      4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.029 n=4+4)
CacheParallel/Get-4      0.00           0.00          ~     (all equal)
```

Note that this uses a temporary fork of https://github.com/hashicorp/golang-lru/pull/111 ... which, cross fingers, will get merged soon.